backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.8.2
pydantic-settings==2.4.0
pandera==0.20.4
pandas==2.2.3
pyarrow==17.0.0
python-dotenv==1.0.1
sse-starlette==2.1.3
apscheduler==3.10.4
anthropic>=0.41.0
scikit-learn==1.5.2
lightgbm==4.5.0
shap>=0.45.0
numpy>=1.26.4
tavily-python==0.5.0
langgraph==0.2.55
langchain-anthropic==0.3.3
